Unit roles - initial semi-stable merge (#187)
* initial commit - stats not loading yet

* fix new stats loading

* basic roles system

* role mechanics for dogfight

* role exp and promotions, dogfight exp

* pilot role dogfight bonuses and stat improvement

* fix for #138

* pilot role for fta

* science and late game pilots stats

* small fixes to science stats

* bravery display fix

* new research list + fixes

* scroll offset decrease

* soldier duty + research state adjastments

* scientists list

* Update InventoryState.cpp

Загрузила Surface ранга роли солдата.

* FixedCompilationError

* Fixed

* ResearchInfoStateFtA

* intermediate result scientist's role

* uncommenting debugDisplayScript

* research mostly done

* fix for ResearchInfoStateFtA::getStatString

* scientist training prototype

* role stats + scientists promotion

* dont forget to clear the project from soldier

* craft filling on new game respects roles

* Scientist stat information

* some progress in engineer role

* some bugs fixed

* some more engineer role UI

* fixes to engineers list

* fixes to manufacture states

* engineers finilly works

* research entermidiate

* research finally done

* refactoring _ftaUi to state

* Revert "refactoring _ftaUi to state"

This reverts commit cd6e240f3b2ac2cddd8ddb542a5510201a34e21c.

* refactor + initial for facility manufacture

* sort soldiers by role

* ongoing work on facility construction

* a bit to facility construction

* facilities manufacture progress

* some work towards intel projects

* intel poject internal logic

* ongiong intel project

* fixes

* prisoners saving on base

* some work to prisoner actions

* minor polish

* cleanings

* #86 fix

* some work for prisoner actions

* some result

* prisoner info

* prisoners

* prisoner report string generation fix

* production save fixes

* manufacturing facility init fixes

* facility manufacturing complete

* more precise man/hour calculations

* agents for soldier info, some unitStat rework

* _initialStats.psiSkill respects maxStat value

* no global research and production if ftaGame

* soldier rank is not used if fta role rank data presented

* clear base duty on transformation

* covert ops start ui refactor

* covert operations exp refactor

* initial commit - stats not loading yet

* fix new stats loading

* basic roles system

* role mechanics for dogfight

* role exp and promotions, dogfight exp

* pilot role dogfight bonuses and stat improvement

* fix for #138

* pilot role for fta

* science and late game pilots stats

* small fixes to science stats

* bravery display fix

* new research list + fixes

* scroll offset decrease

* soldier duty + research state adjastments

* scientists list

* Update InventoryState.cpp

Загрузила Surface ранга роли солдата.

* ResearchInfoStateFtA

* FixedCompilationError

* Fixed

* intermediate result scientist's role

* uncommenting debugDisplayScript

* research mostly done

* fix for ResearchInfoStateFtA::getStatString

* scientist training prototype

* role stats + scientists promotion

* dont forget to clear the project from soldier

* craft filling on new game respects roles

* some progress in engineer role

* Scientist stat information

* some bugs fixed

* some more engineer role UI

* fixes to manufacture states

* fixes to engineers list

* engineers finilly works

* research entermidiate

* research finally done

* refactoring _ftaUi to state

* Revert "refactoring _ftaUi to state"

This reverts commit cd6e240f3b2ac2cddd8ddb542a5510201a34e21c.

* refactor + initial for facility manufacture

* sort soldiers by role

* ongoing work on facility construction

* a bit to facility construction

* facilities manufacture progress

* some work towards intel projects

* intel poject internal logic

* ongiong intel project

* fixes

* prisoners saving on base

* some work to prisoner actions

* minor polish

* cleanings

* #86 fix

* some work for prisoner actions

* prisoner info

* some result

* prisoners

* prisoner report string generation fix

* production save fixes

* manufacturing facility init fixes

* facility manufacturing complete

* more precise man/hour calculations

* agents for soldier info, some unitStat rework

* _initialStats.psiSkill respects maxStat value

* no global research and production if ftaGame

* soldier rank is not used if fta role rank data presented

* clear base duty on transformation

* covert ops start ui refactor

* covert operations exp refactor

* post-rebase fixes

* build errors fixes

* build fixes pt2

* build fix pt 3

* build fix pt 4

* build errors and warnings fix pt 5

* build errors fix pt 6

+++ src/Mod/RuleDiplomacyFaction.cpp
@@ -61,7 +61,7 @@ void RuleDiplomacyFaction::load(const YAML::Node &node)
 	_buyPriceFactor = node["buyPriceFactor"].as<int>(_buyPriceFactor);
 	_repPriceFactor = node["repPriceFactor"].as<int>(_repPriceFactor);
 	_stockMod = node["stockMod"].as<int>(_stockMod);
-	_wishList = node["wishList"].as<std::map<std::string, double>>(_wishList);
+	_wishList = node["wishList"].as<std::map<std::string, int>>(_wishList);
 	_staffWeights = node["staffWeights"].as<std::map<std::string, int>>(_staffWeights);
 	_powerHungry = node["powerHungry"].as<int>(_powerHungry);
 	_scienceBaseCost = node["scienceBaseCost"].as<int>(_scienceBaseCost);
